intermediate/poc-06-rag-multi-document-types/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
openai>=1.40.0
faiss-cpu>=1.8.0
pypdf>=4.3.0
python-docx>=1.1.2
pytest>=8.3.0
httpx>=0.27.0
